出力画像の解像度を向上

+++ OutputRange.vba
@@ -28,8 +28,12 @@ Sub OutputRangeAllSheet()
     Next i
     
     ' 画像出力時の中間ファイルを削除
-    FSO.DeleteFolder (sOutputFolderPath + "\image.files")
-    FSO.DeleteFile (sOutputFolderPath + "\image.htm")
+    If FSO.FolderExists(sOutputFolderPath + "\image.files") Then
+        FSO.DeleteFolder (sOutputFolderPath + "\image.files")
+    End If
+    If FSO.FileExists(sOutputFolderPath + "\image.htm") Then
+        FSO.DeleteFile (sOutputFolderPath + "\image.htm")
+    End If
     
     ' 最後に1シート目を開いてから終了することで、次回起動時に1シート目が表示されるようにする
     Worksheets(1).Activate
@@ -48,6 +52,9 @@ Sub OutputRange(ByVal sSheetName As String, ByVal sRange As String)
     ' 処理対象シートをアクティブにする
     Worksheets(sSheetName).Activate
     
+    ' 処理対象シートのA1セルを選択する
+    Worksheets(sSheetName).Range("A1").Select
+    
     ' 出力先フォルダが存在しなければエラー終了
     sOutputFolderPath = ActiveWorkbook.Path + "\OutputRange"
     If Dir(sOutputFolderPath, vbDirectory) = "" Then
@@ -58,7 +65,7 @@ Sub OutputRange(ByVal sSheetName As String, ByVal sRange As String)
     Application.DisplayAlerts = False
         
     ' 選択範囲を画像としてExcel上に貼り付ける
-    Range(sRange).CopyPicture Appearance:=xlScreen, Format:=xlBitmap
+    Range(sRange).CopyPicture Appearance:=xlScreen, Format:=xlPicture
     ActiveSheet.Paste
     
     ' Excel上に貼り付けた画像をファイル出力(1次出力)
@@ -68,7 +75,7 @@ Sub OutputRange(ByVal sSheetName As String, ByVal sRange As String)
         .AutoRepublish = False
     End With
 
-    ' 1次出力先フォルダから、最終出力先にファイルをコピー
+    ' 1次出力先フォルダから最終出力先にファイルをコピー(シート内に図形が含まれる場合は001を002などに変更する必要あり)
     FileCopy _
         Source:=sOutputFolderPath + "\image.files\AAA_image001.png", _
         Destination:=sOutputFolderPath + "\" + ActiveSheet.Name + ".png"
